combine individual compiler extension classes
